pyproject.toml
[project]
name = "snippet-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "fastapi[standard]>=0.115.6",
    "jinja2>=3.1.5",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.6",
]
